backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0

# ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0

# ONNX export
onnx>=1.16.0